Terrain: added the fk check in delete, added test

+++ src/main/java/com/mcommings/campaigner/services/TerrainService.java
@@ -3,23 +3,33 @@ package com.mcommings.campaigner.services;
 import com.mcommings.campaigner.interfaces.ITerrain;
 import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.Terrain;
+import com.mcommings.campaigner.models.repositories.IPlaceRepository;
 import com.mcommings.campaigner.models.repositories.ITerrainRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_TERRAIN;
 
 @Service
 public class TerrainService implements ITerrain {
 
     private final ITerrainRepository terrainRepository;
+    private final IPlaceRepository placeRepository;
 
     @Autowired
-    public TerrainService(ITerrainRepository terrainRepository) {this.terrainRepository = terrainRepository;}
+    public TerrainService(ITerrainRepository terrainRepository, IPlaceRepository placeRepository) {
+        this.terrainRepository = terrainRepository;
+        this.placeRepository = placeRepository;
+    }
+
     @Override
     public List<Terrain> getTerrains() {
         return terrainRepository.findAll();
@@ -44,7 +54,9 @@ public class TerrainService implements ITerrain {
         if (RepositoryHelper.cannotFindId(terrainRepository, terrainId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-        //TODO: check if foreign key
+        if (RepositoryHelper.isForeignKey(getReposWhereTerrainIsAForeignKey(), FK_TERRAIN.columnName, terrainId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
 
         terrainRepository.deleteById(terrainId);
     }
@@ -59,5 +71,10 @@ public class TerrainService implements ITerrain {
         Terrain terrainToUpdate = RepositoryHelper.getById(terrainRepository, terrainId);
         terrainToUpdate.setName(terrain.getName());
         terrainToUpdate.setDescription(terrain.getDescription());
+    }
+
+    private List<CrudRepository> getReposWhereTerrainIsAForeignKey() {
+        List<CrudRepository> repositories = new ArrayList<>(Arrays.asList(placeRepository));
+        return repositories;
     }
 }
